requirements_detection.txt
# Requirements for Indonesian Car Detection System
# Computer Vision and Deep Learning
opencv-python==4.8.1.78
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196

# Image Processing
Pillow==10.0.0
scikit-image==0.21.0
imageio==2.31.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6

# Utilities
requests==2.31.0
tqdm==4.65.0
pyyaml==6.0.1

# Object Detection Specific
coco-tools==1.1.4
pycocotools==2.0.7

# Performance
numba==0.57.1

# Development
jupyter==1.0.0
notebook==7.0.2
ipywidgets==8.0.7

